add all the models
