+++ src/button/button.model.ts
@@ -1,0 +1,8 @@
+import { Dynamic, Intent, SharedProps, Variant } from '../../../models';
+
+export type ButtonProps = {
+  variant?: Dynamic<Variant>;
+  intent?: Dynamic<Intent>;
+  outline?: boolean;
+  disabled?: boolean;
+} & SharedProps;
