+++ src/column/column.lite.tsx
@@ -1,10 +1,7 @@
 import { onMount, useMetadata, useStore } from '@builder.io/mitosis';
 import { classesToString, getBreakpointClasses } from '../../../helpers';
-import { BreakpointProps, SharedProps } from '../../../models';
 import './column.css';
-
-export type ColumnProps = SharedProps &
-  BreakpointProps<1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 'content' | 'fill' | 'hide'>;
+import { ColumnProps } from './column.model';
 
 useMetadata({ isAttachedToShadowDom: true });
