+++ src/container/container.model.ts
@@ -1,0 +1,5 @@
+import { SharedProps } from '../../../models';
+
+export type ContainerProps = {
+  fluid?: boolean;
+} & SharedProps;
