+++ src/itchio/itchio.model.ts
@@ -1,0 +1,10 @@
+import { SharedProps } from '../../../models';
+
+export type ItchioProps = {
+  user: string;
+  game: string;
+  width?: number;
+  height?: number;
+  onLoad?: () => void;
+  secret?: string;
+} & SharedProps;
