+++ src/pill/pill.model.ts
@@ -1,0 +1,7 @@
+import { Dynamic, Intent, SharedProps, Variant } from '../../../models';
+
+export type PillProps = {
+  variant?: Dynamic<Variant>;
+  intent?: Dynamic<Intent>;
+  disabled?: boolean;
+} & SharedProps;
