+++ src/row/row.model.ts
@@ -1,0 +1,3 @@
+import { BreakpointProps, SharedProps } from '../../../models';
+
+export type RowProps = SharedProps & BreakpointProps<'row' | 'column' | 'row-reverse' | 'column-reverse'>;
