+++ src/spinner/spinner.model.ts
@@ -1,0 +1,7 @@
+import { Dynamic, SharedProps, Variant } from '../../../models';
+
+export type SpinnerProps = {
+  variant?: Dynamic<Variant>;
+  full?: boolean;
+  fullscreen?: boolean;
+} & SharedProps;
